move load/store stuff into memory.rs module
