+++ crates/core/src/value.rs
@@ -38,12 +38,6 @@ impl ValType {
     }
 }
 
-/// Convert one type to another by wrapping.
-pub trait WrapInto<T> {
-    /// Convert one type to another by wrapping.
-    fn wrap_into(self) -> T;
-}
-
 /// Convert one type to another by rounding to the nearest integer towards zero.
 ///
 /// # Errors
@@ -77,96 +71,11 @@ pub trait TruncateSaturateInto<T> {
     fn truncate_saturate_into(self) -> T;
 }
 
-/// Convert one type to another by extending with leading zeroes.
-pub trait ExtendInto<T> {
-    /// Convert one type to another by extending with leading zeroes.
-    fn extend_into(self) -> T;
-}
-
 /// Sign-extends `Self` integer type from `T` integer type.
 pub trait SignExtendFrom<T> {
     /// Convert one type to another by extending with leading zeroes.
     fn sign_extend_from(self) -> Self;
 }
-
-/// Allows to efficiently load bytes from `memory` into a buffer.
-pub trait LoadInto {
-    /// Loads bytes from `memory` into `self`.
-    ///
-    /// # Errors
-    ///
-    /// Traps if the `memory` access is out of bounds.
-    fn load_into(&mut self, memory: &[u8], address: usize) -> Result<(), TrapCode>;
-}
-
-impl<const N: usize> LoadInto for [u8; N] {
-    #[inline]
-    fn load_into(&mut self, memory: &[u8], address: usize) -> Result<(), TrapCode> {
-        let slice: &Self = memory
-            .get(address..)
-            .and_then(|slice| slice.get(..N))
-            .and_then(|slice| slice.try_into().ok())
-            .ok_or(TrapCode::MemoryOutOfBounds)?;
-        *self = *slice;
-        Ok(())
-    }
-}
-
-/// Allows to efficiently write bytes from a buffer into `memory`.
-pub trait StoreFrom {
-    /// Writes bytes from `self` to `memory`.
-    ///
-    /// # Errors
-    ///
-    /// Traps if the `memory` access is out of bounds.
-    fn store_from(&self, memory: &mut [u8], address: usize) -> Result<(), TrapCode>;
-}
-
-impl<const N: usize> StoreFrom for [u8; N] {
-    #[inline]
-    fn store_from(&self, memory: &mut [u8], address: usize) -> Result<(), TrapCode> {
-        let slice: &mut Self = memory
-            .get_mut(address..)
-            .and_then(|slice| slice.get_mut(..N))
-            .and_then(|slice| slice.try_into().ok())
-            .ok_or(TrapCode::MemoryOutOfBounds)?;
-        *slice = *self;
-        Ok(())
-    }
-}
-
-/// Types that can be converted from and to little endian bytes.
-pub trait LittleEndianConvert {
-    /// The little endian bytes representation.
-    type Bytes: Default + LoadInto + StoreFrom;
-
-    /// Converts `self` into little endian bytes.
-    fn into_le_bytes(self) -> Self::Bytes;
-
-    /// Converts little endian bytes into `Self`.
-    fn from_le_bytes(bytes: Self::Bytes) -> Self;
-}
-
-macro_rules! impl_little_endian_convert_primitive {
-    ( $($primitive:ty),* $(,)? ) => {
-        $(
-            impl LittleEndianConvert for $primitive {
-                type Bytes = [::core::primitive::u8; ::core::mem::size_of::<$primitive>()];
-
-                #[inline]
-                fn into_le_bytes(self) -> Self::Bytes {
-                    <$primitive>::to_le_bytes(self)
-                }
-
-                #[inline]
-                fn from_le_bytes(bytes: Self::Bytes) -> Self {
-                    <$primitive>::from_le_bytes(bytes)
-                }
-            }
-        )*
-    };
-}
-impl_little_endian_convert_primitive!(u8, u16, u32, u64, i8, i16, i32, i64, f32, f64);
 
 /// Integer value.
 pub trait Integer: Sized + Unsigned {
@@ -262,26 +171,6 @@ pub trait Float: Sized {
     fn copysign(lhs: Self, rhs: Self) -> Self;
 }
 
-macro_rules! impl_wrap_into {
-    ($from:ident, $into:ident) => {
-        impl WrapInto<$into> for $from {
-            #[inline]
-            fn wrap_into(self) -> $into {
-                self as $into
-            }
-        }
-    };
-}
-
-impl_wrap_into!(i32, i8);
-impl_wrap_into!(i32, i16);
-impl_wrap_into!(i64, i8);
-impl_wrap_into!(i64, i16);
-impl_wrap_into!(i64, i32);
-
-impl_wrap_into!(u32, u32);
-impl_wrap_into!(u64, u64);
-
 macro_rules! impl_try_truncate_into {
     (@primitive $from: ident, $into: ident, $rmin:literal, $rmax:literal) => {
         impl TryTruncateInto<$into, TrapCode> for $from {
@@ -323,34 +212,6 @@ impl_try_truncate_into!(@primitive f32, i64, -9223373136366403584.0_f32,  922337
 impl_try_truncate_into!(@primitive f32, u64,                   -1.0_f32, 18446744073709551616.0_f32);
 impl_try_truncate_into!(@primitive f64, i64, -9223372036854777856.0_f64,  9223372036854775808.0_f64);
 impl_try_truncate_into!(@primitive f64, u64,                   -1.0_f64, 18446744073709551616.0_f64);
-
-macro_rules! impl_extend_into {
-    ($from:ident, $into:ident) => {
-        impl ExtendInto<$into> for $from {
-            #[inline]
-            #[allow(clippy::cast_lossless)]
-            fn extend_into(self) -> $into {
-                self as $into
-            }
-        }
-    };
-}
-
-impl_extend_into!(i8, i32);
-impl_extend_into!(u8, i32);
-impl_extend_into!(i16, i32);
-impl_extend_into!(u16, i32);
-impl_extend_into!(i8, i64);
-impl_extend_into!(u8, i64);
-impl_extend_into!(i16, i64);
-impl_extend_into!(u16, i64);
-impl_extend_into!(i32, i64);
-impl_extend_into!(u32, i64);
-impl_extend_into!(u32, u64);
-
-// Casting to self
-impl_extend_into!(u32, u32);
-impl_extend_into!(u64, u64);
 
 macro_rules! impl_sign_extend_from {
     ( $( impl SignExtendFrom<$from_type:ty> for $for_type:ty; )* ) => {
