+++ crates/core/src/wasm.rs
@@ -1,16 +1,13 @@
 //! Execution helpers for Wasm or Wasmi instructions.
 
 use crate::{
-    value::{LoadInto, StoreFrom},
-    ExtendInto,
+    memory::{load_extend, load_extend_at, store_wrap, store_wrap_at},
     Float,
     Integer,
-    LittleEndianConvert,
     SignExtendFrom,
     TrapCode,
     TruncateSaturateInto,
     TryTruncateInto,
-    WrapInto,
 };
 use core::ops::Neg;
 
@@ -286,47 +283,6 @@ impl_untyped_val! {
     fn i64_trunc_sat_f64_u(value: f64) -> u64 = TruncateSaturateInto::truncate_saturate_into;
 }
 
-/// Calculates the effective address of a linear memory access.
-///
-/// # Errors
-///
-/// If the resulting effective address overflows.
-fn effective_address(ptr: u64, offset: u64) -> Result<usize, TrapCode> {
-    let Some(address) = ptr.checked_add(offset) else {
-        return Err(TrapCode::MemoryOutOfBounds);
-    };
-    usize::try_from(address).map_err(|_| TrapCode::MemoryOutOfBounds)
-}
-
-/// Executes a generic `T.loadN_[s|u]` Wasm operation.
-///
-/// # Errors
-///
-/// - If `ptr + offset` overflows.
-/// - If `ptr + offset` loads out of bounds from `memory`.
-fn load_extend<T, U>(memory: &[u8], ptr: u64, offset: u64) -> Result<T, TrapCode>
-where
-    U: LittleEndianConvert + ExtendInto<T>,
-{
-    let address = effective_address(ptr, offset)?;
-    load_extend_at::<T, U>(memory, address)
-}
-
-/// Executes a generic `T.loadN_[s|u]` Wasm operation.
-///
-/// # Errors
-///
-/// If `address` loads out of bounds from `memory`.
-fn load_extend_at<T, U>(memory: &[u8], address: usize) -> Result<T, TrapCode>
-where
-    U: LittleEndianConvert + ExtendInto<T>,
-{
-    let mut buffer = <<U as LittleEndianConvert>::Bytes as Default>::default();
-    buffer.load_into(memory, address)?;
-    let value: T = <U as LittleEndianConvert>::from_le_bytes(buffer).extend_into();
-    Ok(value)
-}
-
 macro_rules! gen_load_fn {
     (
         (fn $load_fn:ident, fn $load_at_fn:ident, $ty:ty); $($rest:tt)*
@@ -376,37 +332,6 @@ gen_load_fn! {
     (fn i64_load16_u, fn i64_load16_u_at, u16 => i64);
     (fn i64_load32_s, fn i64_load32_s_at, i32 => i64);
     (fn i64_load32_u, fn i64_load32_u_at, u32 => i64);
-}
-
-/// Executes a generic `T.store[N]` Wasm operation.
-///
-/// # Errors
-///
-/// - If `ptr + offset` overflows.
-/// - If `ptr + offset` stores out of bounds from `memory`.
-fn store_wrap<T, U>(memory: &mut [u8], ptr: u64, offset: u64, value: T) -> Result<(), TrapCode>
-where
-    T: WrapInto<U>,
-    U: LittleEndianConvert,
-{
-    let address = effective_address(ptr, offset)?;
-    store_wrap_at::<T, U>(memory, address, value)
-}
-
-/// Executes a generic `T.store[N]` Wasm operation.
-///
-/// # Errors
-///
-/// - If `address` stores out of bounds from `memory`.
-fn store_wrap_at<T, U>(memory: &mut [u8], address: usize, value: T) -> Result<(), TrapCode>
-where
-    T: WrapInto<U>,
-    U: LittleEndianConvert,
-{
-    let wrapped = value.wrap_into();
-    let buffer = <U as LittleEndianConvert>::into_le_bytes(wrapped);
-    buffer.store_from(memory, address)?;
-    Ok(())
 }
 
 macro_rules! gen_store_fn {
